Centralize StringBuilderCache
